+++ previous-versions/banking/pkg/models/operations/listtransactioncategories.go
@@ -79,8 +79,6 @@ func (o *ListTransactionCategoriesRequest) GetQuery() *string {
 type ListTransactionCategoriesResponse struct {
 	// HTTP response content type for this operation
 	ContentType string
-	// Your `query` parameter was not correctly formed
-	ErrorMessage *shared.ErrorMessage
 	// HTTP response status code for this operation
 	StatusCode int
 	// Raw HTTP response; suitable for custom response parsing
@@ -94,13 +92,6 @@ func (o *ListTransactionCategoriesResponse) GetContentType() string {
 		return ""
 	}
 	return o.ContentType
-}
-
-func (o *ListTransactionCategoriesResponse) GetErrorMessage() *shared.ErrorMessage {
-	if o == nil {
-		return nil
-	}
-	return o.ErrorMessage
 }
 
 func (o *ListTransactionCategoriesResponse) GetStatusCode() int {